RAD-63: Remove duplicate call to setupCrashHandler
Removed duplicate call to setupCrashHandler, which should only be called
once for the application. It probably won't break anything to call it twice
but it shouldn't be called in place of exception handling.

An UnhandledExceptionHandler is used to catch exceptions that aren't caught
in any try/catch block. If an exception is caught, then it must be handled
appropriately. If there is no appropriate way to handle the exception, then
it should thrown and not caught. Note: this is tricky for any exceptions that
don't extend RuntimeException because they must be caught.

Also removed some code that was no longer required and was resulting in
an exception that needed handling.