RAD-63: Fix error message handling
Fixed error message handling which was an issue when an exception occurred.
Because the exception is being caught, the onPostExecute method is always
called. (If an exception was thrown from the method, then it would crash
the app or be handled by the UncaughtExceptionHandler and the onPostExecute
method would not be invoked.)

Therefore, the AsyncTask should return a boolean value indicating if it
was successful or not, and show the correct Toast message.