RAD-74: Fix bug in deserialization code
Fixed bug in deserialization code where the toIndex was being used instead
of the fromIndex to determine the next delimiter.
Also added comments about the various exceptions that can be thrown when
processing the bytes. We need to consider how to handle these so we are
able to read cards that may be partially corrupted.